requirements.txt
openai
pinecone-client
langchain==0.0.266